✨ Create representations of Category-related classes in API module
The main classes were extracted in representations and moved to the
API module. Some refactoring and reference updates were needed.
