+++ features/category-api/build.gradle
@@ -1,1 +1,8 @@
 apply from: "$rootDir/config/dependencies/feature_dependencies.gradle"
+apply from: "$rootDir/config/dependencies/compose_dependencies.gradle"
+
+dependencies {
+    implementation project(":data:domain")
+
+    implementation Deps.koin.android
+}
