+++ features/category-api/src/main/java/com/escodro/categoryapi/presentation/CategoryState.kt
@@ -1,0 +1,26 @@
+package com.escodro.categoryapi.presentation
+
+import com.escodro.categoryapi.model.Category
+
+/**
+ * Represents the states of [CategoryListViewModel].
+ */
+sealed class CategoryState {
+
+    /**
+     * Loading state.
+     */
+    object Loading : CategoryState()
+
+    /**
+     * Loaded state.
+     *
+     * @param categoryList the loaded category list
+     */
+    data class Loaded(val categoryList: List<Category>) : CategoryState()
+
+    /**
+     * Empty state, there are no categories to be shown.
+     */
+    object Empty : CategoryState()
+}
