+++ features/category/build.gradle
@@ -13,6 +13,7 @@ android {
 }
 
 dependencies {
+    implementation project(":features:category-api")
     implementation project(":data:domain")
     implementation project(":libraries:theme")
