+++ features/task/build.gradle
@@ -14,6 +14,7 @@ android {
 }
 
 dependencies {
+    implementation project(":features:category-api")
     implementation project(":libraries:core")
     implementation project(":data:domain")
     implementation project(":libraries:theme")
