+++ features/task/src/androidTest/java/com/escodro/task/CategorySelectionTest.kt
@@ -10,9 +10,9 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.test.platform.app.InstrumentationRegistry
-import com.escodro.task.model.Category
+import com.escodro.categoryapi.model.Category
+import com.escodro.categoryapi.presentation.CategoryState
 import com.escodro.task.presentation.category.CategorySelection
-import com.escodro.task.presentation.category.CategoryState
 import com.escodro.task.presentation.category.ChipNameKey
 import com.escodro.theme.AlkaaTheme
 import org.junit.Ignore
