+++ features/task/src/main/java/com/escodro/task/mapper/TaskWithCategoryMapper.kt
@@ -1,5 +1,6 @@
 package com.escodro.task.mapper
 
+import com.escodro.categoryapi.mapper.CategoryMapper
 import com.escodro.domain.model.TaskWithCategory as DomainTaskWithCategory
 import com.escodro.task.model.TaskWithCategory as ViewTaskWithCategory
