--- features/task/src/main/java/com/escodro/task/presentation/category/CategoryState.kt
@@ -1,12 +1,0 @@
-package com.escodro.task.presentation.category
-
-import com.escodro.task.model.Category
-
-internal sealed class CategoryState {
-
-    object Loading : CategoryState()
-
-    data class Loaded(val categoryList: List<Category>) : CategoryState()
-
-    object Empty : CategoryState()
-}
